+++ 2024/lib/hashtable.h
@@ -3,6 +3,7 @@
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct Hashtable Hashtable;
 
@@ -11,6 +12,6 @@ void hashtable_free(Hashtable *ht);
 bool hashtable_has(Hashtable *ht, const char *key);
 void *hashtable_get(Hashtable *ht, const char *key);
 Hashtable *hashtable_set(Hashtable *ht, const char *key, const void *value);
-size_t hashtable_size(Hashtable *ht);
+int64_t hashtable_size(Hashtable *ht);
 
 #endif
