+++ 2024/lib/txt.c
@@ -3,16 +3,17 @@
 #include <assert.h>
 #include <errno.h>
 #include <stddef.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
 #include "txt.h"
 
-Txt *txt_new(size_t num_lines) {
+Txt *txt_new(int64_t num_lines) {
   Txt *txt = malloc(sizeof(*txt));
   assert(txt != NULL && "malloc failed");
 
-  txt->lines = malloc(sizeof(*txt->lines) * num_lines);
+  txt->lines = malloc(sizeof(*txt->lines) * (size_t)num_lines);
   assert(txt->lines != NULL && "malloc failed");
 
   txt->num_lines = num_lines;
@@ -28,7 +29,7 @@ Txt *txt_read(FILE *stream) {
   errno = 0;
   while (getline(&line, &buf_len, stream) != -1) {
     txt->lines =
-        realloc(txt->lines, sizeof(*txt->lines) * (txt->num_lines + 1));
+        realloc(txt->lines, sizeof(*txt->lines) * ((size_t)txt->num_lines + 1));
     assert(txt->lines != NULL && "realloc failed");
 
     line[strcspn(line, "\n")] = '\0';
@@ -42,7 +43,7 @@ Txt *txt_read(FILE *stream) {
 }
 
 void txt_free(Txt *txt) {
-  for (size_t i = 0; i < txt->num_lines; i++) {
+  for (int64_t i = 0; i < txt->num_lines; i++) {
     free(txt->lines[i]);
   }
   free(txt->lines);
