+++ 2024/lib/txt.h
@@ -2,14 +2,15 @@
 #define TXT_H
 
 #include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 
 typedef struct Txt {
   char **lines;
-  size_t num_lines;
+  int64_t num_lines;
 } Txt;
 
-Txt *txt_new(size_t num_lines);
+Txt *txt_new(int64_t num_lines);
 Txt *txt_read(FILE *stream);
 void txt_free(Txt *txt);
